+++ pipelinewise/fastsync/commons/tap_mysql.py
@@ -1,10 +1,13 @@
 import csv
 import datetime
 import decimal
+import glob
 import logging
+import os
 import pymysql
 import pymysql.cursors
 
+from argparse import Namespace
 from typing import Tuple, Dict, Callable
 from pymysql import InterfaceError, OperationalError, Connection
 
@@ -277,11 +280,11 @@ class FastSyncTapMySql:
         Get the actual incremental key position in the table
         """
         result = self.query(
-            'SELECT MAX({}) AS key_value FROM {}'.format(replication_key, table)
+            f'SELECT MAX({replication_key}) AS key_value FROM {table}'
         )
         if not result:
             raise Exception(
-                'Cannot get replication key value for table: {}'.format(table)
+                f'Cannot get replication key value for table: {table}'
             )
 
         mysql_key_value = result[0].get('key_value')
@@ -308,9 +311,8 @@ class FastSyncTapMySql:
         Get the primary key of a table
         """
         table_dict = utils.tablename_to_dict(table_name)
-        sql = "SHOW KEYS FROM `{}`.`{}` WHERE Key_name = 'PRIMARY'".format(
-            table_dict['schema_name'], table_dict['table_name']
-        )
+        sql = f"SHOW KEYS FROM `{table_dict['schema_name']}`.`{table_dict['table_name']}` WHERE Key_name = 'PRIMARY'"
+
         pk_specs = self.query(sql)
         if len(pk_specs) > 0:
             return [
@@ -414,6 +416,7 @@ class FastSyncTapMySql:
             split_file_chunk_size_mb=1000,
             split_file_max_chunks=20,
             compress=True,
+            where_clause_sql='',
     ):
         """
         Export data from table to a zipped csv
@@ -433,15 +436,20 @@ class FastSyncTapMySql:
             raise Exception('{} table not found.'.format(table_name))
 
         table_dict = utils.tablename_to_dict(table_name)
+
+        column_safe_sql_values = column_safe_sql_values + [
+            "CONVERT_TZ( NOW(),@@session.time_zone,'+00:00') AS `_SDC_EXTRACTED_AT`",
+            "CONVERT_TZ( NOW(),@@session.time_zone,'+00:00') AS `_SDC_BATCHED_AT`",
+            'null AS `_SDC_DELETED_AT`'
+        ]
+
         sql = """SELECT {}
-        ,CONVERT_TZ( NOW(),@@session.time_zone,'+00:00') AS _SDC_EXTRACTED_AT
-        ,CONVERT_TZ( NOW(),@@session.time_zone,'+00:00') AS _SDC_BATCHED_AT
-        ,null AS _SDC_DELETED_AT
-        FROM `{}`.`{}`
+        FROM `{}`.`{}` {}
         """.format(
             ','.join(column_safe_sql_values),
             table_dict['schema_name'],
             table_dict['table_name'],
+            where_clause_sql
         )
         export_batch_rows = self.connection_config['export_batch_rows']
         exported_rows = 0
@@ -487,6 +495,25 @@ class FastSyncTapMySql:
                 LOGGER.info(
                     'Exported total of %s rows from %s...', exported_rows, table_name
                 )
+
+    def export_source_table_data(
+            self, args: Namespace, tap_id: str, where_clause_sql: str = '') -> list:
+        """Export source table data"""
+        filename = utils.gen_export_filename(tap_id=tap_id, table=args.table, sync_type='partialsync')
+        filepath = os.path.join(args.temp_dir, filename)
+
+        # Exporting table data
+
+        self.copy_table(
+            args.table,
+            filepath,
+            split_large_files=args.target.get('split_large_files'),
+            split_file_chunk_size_mb=args.target.get('split_file_chunk_size_mb'),
+            split_file_max_chunks=args.target.get('split_file_max_chunks'),
+            where_clause_sql=where_clause_sql,
+        )
+        file_parts = glob.glob(f'{filepath}*')
+        return file_parts
 
     def __get_primary_server_uuid(self) -> str:
         """